Rename custom.js -> slack.js
