+++ server/api/controllers/cards/delete.js
@@ -1,4 +1,4 @@
-const services = require('../../services/custom');
+const services = require('../../services/slack');
 
 const Errors = {
   NOT_ENOUGH_RIGHTS: {
